refactor: Controller 패키지 v1 통일화
