+++ src/main/java/net/detalk/api/admin/controller/v1/response/GetVisitorLogResponse.java
@@ -1,4 +1,4 @@
-package net.detalk.api.admin.controller.response;
+package net.detalk.api.admin.controller.v1.response;
 
 import java.time.LocalDateTime;
