+++ src/main/java/net/detalk/api/post/repository/impl/ProductPostRepositoryImpl.java
@@ -24,8 +24,8 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import net.detalk.api.post.controller.response.GetProductPostResponse;
-import net.detalk.api.post.controller.response.GetProductPostResponse.Media;
+import net.detalk.api.post.controller.v1.response.GetProductPostResponse;
+import net.detalk.api.post.controller.v1.response.GetProductPostResponse.Media;
 import net.detalk.api.post.domain.ProductPost;
 import net.detalk.api.post.repository.ProductPostRepository;
 import org.jooq.Condition;
