+++ src/main/java/net/detalk/api/post/service/ProductPostService.java
@@ -8,8 +8,8 @@ import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.detalk.api.post.controller.request.UpdateProductPostRequest;
-import net.detalk.api.post.controller.response.GetProductPostResponse;
+import net.detalk.api.post.controller.v1.request.UpdateProductPostRequest;
+import net.detalk.api.post.controller.v1.response.GetProductPostResponse;
 import net.detalk.api.plan.service.PricingPlanService;
 import net.detalk.api.post.repository.ProductPostLastSnapshotRepository;
 import net.detalk.api.post.repository.ProductPostLinkRepository;
@@ -33,7 +33,7 @@ import net.detalk.api.product.domain.ProductMaker;
 import net.detalk.api.post.domain.ProductPostSnapshotAttachmentFile;
 import net.detalk.api.plan.domain.PricingPlan;
 import net.detalk.api.product.domain.Product;
-import net.detalk.api.post.controller.request.CreateProductPostRequest;
+import net.detalk.api.post.controller.v1.request.CreateProductPostRequest;
 import net.detalk.api.post.domain.ProductPost;
 import net.detalk.api.post.domain.ProductPostSnapshot;
 import net.detalk.api.post.domain.ProductPostSnapshotTag;
