+++ src/main/java/net/detalk/api/tag/controller/v1/response/GetTagResponse.java
@@ -1,4 +1,4 @@
-package net.detalk.api.tag.controller.response;
+package net.detalk.api.tag.controller.v1.response;
 
 /**
  * 태그 조회 응답 DTO
