+++ src/test/java/net/detalk/api/service/ProductPostServiceTest.java
@@ -13,10 +13,10 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import net.detalk.api.post.controller.request.CreateProductPostRequest;
-import net.detalk.api.post.controller.request.UpdateProductPostRequest;
-import net.detalk.api.post.controller.response.GetProductPostResponse;
-import net.detalk.api.post.controller.response.GetProductPostResponse.Media;
+import net.detalk.api.post.controller.v1.request.CreateProductPostRequest;
+import net.detalk.api.post.controller.v1.request.UpdateProductPostRequest;
+import net.detalk.api.post.controller.v1.response.GetProductPostResponse;
+import net.detalk.api.post.controller.v1.response.GetProductPostResponse.Media;
 import net.detalk.api.plan.domain.PricingPlan;
 import net.detalk.api.plan.service.PricingPlanService;
 import net.detalk.api.post.repository.ProductPostLastSnapshotRepository;
